# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Timeless Love API
  description: Backend API for Timeless Love family social platform
  version: 1.0.0
  contact:
    name: API Support
    email: support@timelesslove.com
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.timelesslove.com
    description: Production server
...
paths:
...
  /api/v1/memories:
    post:
      tags:
        - memories
      summary: Create Memory
      description: Create a new memory with media references
      operationId: createMemory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMemoryRequest'
      responses:
        '201':
          description: Memory created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'

    get:
      tags:
        - memories
      summary: List Memories
      description: List memories for the user's family unit
      operationId: listMemories
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, archived]
      responses:
        '200':
          description: List of memories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemoryResponse'
  /api/v1/memories/{memory_id}:
    get:
      tags:
        - memories
      summary: Get Memory
      description: Get memory details with associated media
      operationId: getMemory
      security:
        - bearerAuth: []
      parameters:
        - name: memory_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Memory details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'

    put:
      tags:
        - memories
      summary: Update Memory
      description: Update memory metadata
      operationId: updateMemory
      security:
        - bearerAuth: []
      parameters:
        - name: memory_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemoryRequest'
      responses:
        '200':
          description: Memory updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      tags:
        - memories
      summary: Delete Memory
      description: Delete a memory (cascades to media)
      operationId: deleteMemory
      security:
        - bearerAuth: []
      parameters:
        - name: memory_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Memory deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/storage/upload-url:
    post:
      tags:
        - storage
      summary: Generate Upload URL
      description: Generate signed URL for uploading media to Supabase Storage
      operationId: generateUploadUrl
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Upload URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/storage/access-url:
    get:
      tags:
        - storage
      summary: Generate Access URL
      description: Generate signed URL for accessing media from Supabase Storage
      operationId: generateAccessUrl
      security:
        - bearerAuth: []
      parameters:
        - name: storage_path
          in: query
          required: true
          schema:
            type: string
        - name: expires_in
          in: query
          schema:
            type: integer
            default: 3600
      responses:
        '200':
          description: Access URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessUrlResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/storage/media/{media_id}/url:
    get:
      tags:
        - storage
      summary: Get Media URL
      description: Get signed URL for a specific media file by media ID
      operationId: getMediaUrl
      security:
        - bearerAuth: []
      parameters:
        - name: media_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: expires_in
          in: query
          schema:
            type: integer
            default: 3600
      responses:
        '200':
          description: Media URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessUrlResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateMemoryRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        memory_date:
          type: string
          format: date
        location:
          type: string
          maxLength: 255
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, published, archived]
          default: draft
        media:
          type: array
          items:
            $ref: '#/components/schemas/MediaReference'
    MediaReference:
      type: object
      required:
        - storage_path
        - file_name
        - mime_type
        - file_size
      properties:
        storage_path:
          type: string
        file_name:
          type: string
        mime_type:
          type: string
        file_size:
          type: integer
          minimum: 1
    UpdateMemoryRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        memory_date:
          type: string
          format: date
        location:
          type: string
          maxLength: 255
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, published, archived]
    MemoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        family_unit_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        memory_date:
          type: string
          format: date
        location:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        modified_by:
          type: string
          format: uuid
        media:
          type: array
          items:
            $ref: '#/components/schemas/MediaResponse'
    MediaResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        memory_id:
          type: string
          format: uuid
        storage_path:
          type: string
        storage_bucket:
          type: string
        file_name:
          type: string
        mime_type:
          type: string
        file_size:
          type: integer
        width:
          type: integer
        height:
          type: integer
        duration:
          type: integer
        thumbnail_path:
          type: string
        processing_status:
          type: string
          enum: [pending, processing, completed, failed]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UploadUrlRequest:
      type: object
      required:
        - memory_id
        - file_name
        - mime_type
      properties:
        memory_id:
          type: string
          format: uuid
        file_name:
          type: string
        mime_type:
          type: string
    UploadUrlResponse:
      type: object
      properties:
        upload_url:
          type: string
          format: uri
        storage_path:
          type: string
        expires_in:
          type: integer
    AccessUrlResponse:
      type: object
      properties:
        access_url:
          type: string
          format: uri
        expires_in:
          type: integer
    Error:
      type: object
      properties:
        detail:
          type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...